started gtp - GNU Telescope Pointing library
