+++ include/gtp/Makefile.am
@@ -1,0 +1,1 @@
+noinst_HEADERS = Vector.h Point.hpp ppoly.h
