+++ lib/Makefile.am
@@ -1,1 +1,16 @@
-SUBDIRS = xmlrpc++ sep rts2 pluto sgp4 rts2fits rts2db rts2script rts2scheduler rts2json libindi rts2tel rts2lx200 ucac5
+SUBDIRS = \
+	xmlrpc++ \
+	sep \
+	rts2 \
+	pluto \
+	sgp4 \
+	rts2fits \
+	rts2db \
+	rts2script \
+	rts2scheduler \
+	rts2json \
+	libindi \
+	rts2tel \
+	rts2lx200 \
+	ucac5 \
+	gtp
