+++ lib/gtp/Makefile.am
@@ -1,0 +1,5 @@
+
+lib_LTLIBRARIES = libgtp.la
+
+libgtp_la_SOURCES = ppoly.cpp
+libgtp_la_CXXFLAGS = -I../../include -I../../include/gtp
